+++ .github/actions/test_update-from-older-version.sh
@@ -20,7 +20,7 @@ if [[ -z $(grep "No migration needed." $LOG_FILE) ]];
   then echo "bin/console glpi:database:update command FAILED" && exit 1;
 fi
 ## Check DB
-bin/console glpi:database:check_schema \
+bin/console glpi:database:check_schema_integrity \
   --config-dir=./tests/config --ansi --no-interaction \
   --ignore-innodb-migration --ignore-timestamps-migration --ignore-dynamic-row-format-migration --ignore-utf8mb4-migration
 
@@ -36,7 +36,7 @@ if [[ -z $(grep "No migration needed." $LOG_FILE) ]];
   then echo "bin/console glpi:migration:myisam_to_innodb command FAILED" && exit 1;
 fi
 ## Check DB
-bin/console glpi:database:check_schema \
+bin/console glpi:database:check_schema_integrity \
   --config-dir=./tests/config --ansi --no-interaction \
   --ignore-timestamps-migration --ignore-dynamic-row-format-migration --ignore-utf8mb4-migration
 
@@ -52,7 +52,7 @@ if [[ -z $(grep "No migration needed." $LOG_FILE) ]];
   then echo "bin/console glpi:migration:timestamps command FAILED" && exit 1;
 fi
 ## Check DB
-bin/console glpi:database:check_schema \
+bin/console glpi:database:check_schema_integrity \
   --config-dir=./tests/config --ansi --no-interaction \
   --ignore-dynamic-row-format-migration --ignore-utf8mb4-migration
 
@@ -60,7 +60,7 @@ bin/console glpi:database:check_schema \
 ## Result will depend on DB server/version, we just expect that command will not fail.
 bin/console glpi:migration:dynamic_row_format --config-dir=./tests/config --ansi --no-interaction
 ## Check DB
-bin/console glpi:database:check_schema --config-dir=./tests/config --ansi --no-interaction --ignore-utf8mb4-migration
+bin/console glpi:database:check_schema_integrity --config-dir=./tests/config --ansi --no-interaction --ignore-utf8mb4-migration
 
 # Execute utf8mb4 migration
 ## First run should do the migration (with no warnings).
@@ -74,7 +74,7 @@ if [[ -z $(grep "No migration needed." $LOG_FILE) ]];
   then echo "bin/console glpi:migration:utf8mb4 command FAILED" && exit 1;
 fi
 ## Check DB
-bin/console glpi:database:check_schema --config-dir=./tests/config --ansi --no-interaction
+bin/console glpi:database:check_schema_integrity --config-dir=./tests/config --ansi --no-interaction
 
 # Check updated data
 bin/console glpi:database:configure \