+++ inc/knowbaseitem_revision.class.php
@@ -172,7 +172,7 @@ class KnowbaseItem_Revision extends CommonDBTM {
          echo "/> <input type='radio' name='diff' value='{$revision['id']}'/></td>";
 
          echo "<td>" . ($hasRevUser ? $user->getLink() : __('Unknown user')) . "</td>".
-             "<td class='tab_date'>". $revision['date_creation'] . "</td>";
+             "<td class='tab_date'>". $revision['date'] . "</td>";
 
          $form = null;
          if ($item->getType() == KnowbaseItem::getType()) {
@@ -299,7 +299,7 @@ class KnowbaseItem_Revision extends CommonDBTM {
       $this->fields['knowbaseitems_id'] = $item->fields['id'];
       $this->fields['name'] = Toolbox::addslashes_deep($item->fields['name']);
       $this->fields['answer'] = Toolbox::addslashes_deep($item->fields['answer']);
-      $this->fields['date_creation'] = $item->fields['date_mod'];
+      $this->fields['date'] = $item->fields['date_mod'];
       $this->fields['revision'] = $this->getNewRevision();
       $this->fields['users_id'] = $item->fields['users_id'];
       $this->addToDB();
@@ -318,7 +318,7 @@ class KnowbaseItem_Revision extends CommonDBTM {
       $this->fields['knowbaseitems_id'] = $item->fields['knowbaseitems_id'];
       $this->fields['name'] = Toolbox::addslashes_deep($item->fields['name']);
       $this->fields['answer'] = Toolbox::addslashes_deep($item->fields['answer']);
-      $this->fields['date_creation'] = $item->fields['date_mod'];
+      $this->fields['date'] = $item->fields['date_mod'];
       $this->fields['language'] = $item->fields['language'];
       $this->fields['revision'] = $this->getNewRevision();
       $this->fields['users_id'] = $item->fields['users_id'];